pyproject.toml
[tool.poetry]
name = "wolt-cli"
version = "1.1.1"
description = "What to eat? CLI tool to interaction with Wolt API"
authors = ["Kamil Woźniak <info@kamilwozniak.com>"]
readme = "README.md"

[tool.poetry.scripts]
wolt-cli = "wolt_cli.main:app"

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.9.0"}
click = ">=8.1.7,<8.2.0"
pydantic = "^2.4.2"
httpx = "^0.25.0"
inquirer = ">=2.10.1,<4.0.0"
pydantic-settings = "^2.0.3"

[tool.poetry.urls]
homepage = "https://github.com/Valaraucoo/wolt-cli"
repository = "https://github.com/Valaraucoo/wolt-cli"
documentation = "https://github.com/Valaraucoo/wolt-cli"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.286"
pytest = "^7.2.2"
mypy = "^1.6.0"
factory-boy = "^3.2.1"
pre-commit = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"
select = ["E", "F", "Q", "U", "N", "C", "B", "T"]
ignore = ["T201", "E999", "B008", "B904", "UP040"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict_optional = false
show_error_codes = true

[tool.mypy-app.cache]
ignore_errors = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
profile = "black"

[tool.black]
line-length = 120
target-version = ["py312"]
